--- AnarchyExploitFixesFolia/src/main/java/me/xginko/aef/modules/combat/AutoBedOrSpigot5988.java
@@ -1,44 +1,0 @@
-package me.xginko.aef.modules.combat;
-
-import com.destroystokyo.paper.event.player.PlayerSetSpawnEvent;
-import me.xginko.aef.modules.AEFModule;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.HandlerList;
-import org.bukkit.event.Listener;
-
-public class AutoBedOrSpigot5988 extends AEFModule implements Listener {
-
-    public AutoBedOrSpigot5988() {
-        super("combat.auto-bed");
-    }
-
-    @Override
-    public void enable() {
-        plugin.getServer().getPluginManager().registerEvents(this, plugin);
-    }
-
-    @Override
-    public void disable() {
-        HandlerList.unregisterAll(this);
-    }
-
-    @Override
-    public boolean shouldEnable() {
-        return config.getBoolean(configPath + ".enable", false, "As it happens, players on a server with a huge game world somehow need to travel long distances.\n" +
-                "From Minecraft version 1.16 (≈June 2020) to version 1.17.1(≈October 2021) in the game there was a bug SPIGOT-5988 which did not reset the respawn point of the player after death, if his bed was blocked.\n" +
-                "After some manipulations, it turned out to be a simple mechanism that allowed the player to get to the spavn by blocking his bed with a shalker before death, and after the second death to be near his bed again.\n" +
-                "This bug persisted from the Spigot server to Paper and all its forks until October 2021, after which it was fixed by the Spigot development team.\n" +
-                "Attempts by players to reach out to Spigot to allow them to disable the patch that fixes the SPIGOT-5988 bug have failed.\n" +
-                "\n" +
-                "Demonstration of how the patch works:\n" +
-                "https://www.youtube.com/watch?v=3y5SbQXzMss&feature=youtu.be");
-    }
-
-    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
-    private void onPlayerSetSpawn(PlayerSetSpawnEvent event) {
-        if (event.getLocation() == null) {
-            event.setCancelled(true);
-        }
-    }
-}
