+++ AnarchyExploitFixesFolia/src/main/java/me/xginko/aef/modules/misc/AutoBedOrSpigot5988.java
@@ -1,0 +1,51 @@
+package me.xginko.aef.modules.misc;
+
+import com.destroystokyo.paper.event.player.PlayerSetSpawnEvent;
+import me.xginko.aef.modules.AEFModule;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.Listener;
+
+public class AutoBedOrSpigot5988 extends AEFModule implements Listener {
+
+    public AutoBedOrSpigot5988() {
+        super("misc.auto-bed");
+        config.addComment(configPath+ ".enable",
+                "Re-enables SPIGOT-5988, also known as 'auto-bed'\n"+
+                        "From Minecraft version 1.16 (≈June 2020) to version 1.17.1(≈October 2021)\n" +
+                        "there was a bug (SPIGOT-5988) which did not reset the respawn point of the\n" +
+                        "player after death, if his bed was blocked with a shulker.\n" +
+                        "After dying a second time, the player will be back at his bed again.\n" +
+                        "This bug persisted from the Spigot server to Paper and all its forks until\n" +
+                        "October 2021, after which it was fixed by the Spigot development team.\n" +
+                        "Attempts by players to reach out to Spigot to allow them to disable the patch\n" +
+                        "that fixes the SPIGOT-5988 bug have failed.\n" +
+                        "Demonstration of how the patch works:\n" +
+                        "https://www.youtube.com/watch?v=3y5SbQXzMss");
+    }
+
+    @Override
+    public void enable() {
+        plugin.getServer().getPluginManager().registerEvents(this, plugin);
+    }
+
+    @Override
+    public void disable() {
+        HandlerList.unregisterAll(this);
+    }
+
+    @Override
+    public boolean shouldEnable() {
+        return config.getBoolean(configPath + ".enable", false);
+    }
+
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
+    private void onPlayerSetSpawn(com.destroystokyo.paper.event.player.PlayerSetSpawnEvent event) {
+        if (event.getCause() == PlayerSetSpawnEvent.Cause.PLUGIN || event.getCause() == PlayerSetSpawnEvent.Cause.COMMAND) return;
+        
+        if (event.getLocation() == null) {
+            event.setCancelled(true);
+        }
+    }
+}
